+++ src/main/java/duke/Command/AddCommand.java
@@ -1,5 +1,9 @@
-package duke;
+package duke.Command;
 
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
 import duke.task.Deadline;
 import duke.task.Event;
 import duke.task.Task;
