+++ src/main/java/duke/Parser.java
@@ -1,7 +1,6 @@
 package duke;
 
-import java.io.IOException;
-import java.util.Scanner;
+import duke.Command.*;
 
 public class Parser {
